+++ misc/sampledata/EMPTY_GROUP/01_central.sql
@@ -1,0 +1,1 @@
+GRANT SELECT, INSERT, UPDATE, DELETE ON TESTTABLE to <region> with grant option;